% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadEAIM.R
\name{CalcMoles}
\alias{CalcMoles}
\title{CalcMoles}
\usage{
CalcMoles(table, vars = c("SO4", "NO3", "NH4", "Cl", "Na"))
}
\arguments{
\item{table}{E-AIM table}

\item{vars}{ions}
}
\value{
table of moles
}
\description{
Calculate moles
}

