% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadEAIM.R
\name{ExpandSolids}
\alias{ExpandSolids}
\title{ExpandSolids}
\usage{
ExpandSolids(table)
}
\arguments{
\item{table}{E-AIM table}
}
\value{
table with expanded solids columns
}
\description{
Tabularize solid columns in E-AIM table
}

