% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTSI.R
\name{merge.TSI}
\alias{merge.TSI}
\title{merge.TSI}
\usage{
\method{merge}{TSI}(x, y, basename = NULL)
}
\arguments{
\item{x}{first TSI object}

\item{y}{second TSI object}
}
\value{
TSI object (list).
}
\description{
Merge TSI objects.
}

