% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadEAIM.R
\name{ReadEAIM}
\alias{ReadEAIM}
\title{ReadEAIM}
\usage{
ReadEAIM(filename, divider = "[-]{5,}")
}
\arguments{
\item{filename}{file name}

\item{divider}{line divider between blocks of data}
}
\value{
List of 3 tables.
}
\description{
Read E-AIM HTML
}

