% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadICT.R
\name{ReadICT1001}
\alias{ReadICT1001}
\title{ReadICT1001}
\usage{
ReadICT1001(filename, format.time = FALSE)
}
\arguments{
\item{filename}{name of file in ICARTT 1001 format}

\item{format.time}{[TRUE/FALSE]}
}
\value{
a data frame with additional attribute, "header", attached. Access header information with \code{header <- attr(dataframe,"header")} or \code{cat(paste(attr(dataframe,"header"),collapse="\n"),"\n")}. Note that modifications to this data frame [e.g., with \code{transform()}] may cause loss of this attribute.
}
\description{
ICARTT data file reader in R
Accepts only format 1001 (comma delimited, all variables are real)
}

