% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadTSI.R
\name{ReadTSI}
\alias{ReadTSI}
\title{ReadTSI}
\usage{
ReadTSI(filename, sep)
}
\arguments{
\item{filename}{file name}

\item{sep}{delimiter}
}
\value{
TSI object (list).
}
\description{
Read output of TSI AIM software
}

