% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteICT.R
\name{JSON2ICT}
\alias{JSON2ICT}
\title{JSON2ICT}
\usage{
JSON2ICT(jsonfile, csvfile, ictfile)
}
\arguments{
\item{jsonfile}{JSON file}

\item{csvfile}{[optional] data file}

\item{ictfile}{[optional] output file name}
}
\value{
produces side effect (writing ICT files)
}
\description{
Generate ICT-formatted files
}

